Подключение PVS-Studio

+++ source/TacticsVictory/src/Core/Math.cpp
@@ -16,7 +16,7 @@ BoundingBox Math::CalculateBoundingBox(float *buffer, uint numVertexes)
 
     for(uint i = 0; i < numVertexes; i++)
     {
-        box.Merge(Vector3(buffer + i * 8));
+        box.Merge(Vector3(buffer + (size_t)(i * 8)));
     }
 
     return box;
